evals/pyproject.toml
[project]
name = "gkg-evals"
version = "0.1.0"
description = "GKG agent evaluation harness"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11",
    "httpx>=0.28",
    "httpx-sse>=0.4",
    "polars>=1.20",
    "duckdb>=1.2",
    "pyyaml>=6.0",
    "click>=8.1",
    "scipy>=1.14",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11",
    "pyright>=1.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]

[project.scripts]
gkg-eval = "harness.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["harness"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
